Update prices_sales and was created par_impar

+++ 01-introduction/price_sales.py
@@ -4,7 +4,7 @@ In this app we will build a software to calculate IVA
 
 value_product = input("Type product value:")
 
-value_product = int(value_product)
+value_product = float(value_product)
 iva = value_product * 18 / 100
 final_value = iva + value_product
